changes according to pr comments

+++ backend/OurFootprint/scripts/dels.py
@@ -2,18 +2,23 @@ from calculator.models import FortisBillField, HydroBillField, Commute, UserComm
 
 
 def del_fortis(uid, fortis_id):
-    instance = FortisBillField.objects.get(id=fortis_id, user_id=uid)
+    instance = FortisBillField.objects.get(id=fortis_id)
     if instance is None:
-        return {"error": "not found"}, 404
+        return {"error": "Not found"}, 404
+    if instance.user_id != uid:
+        return {"error": "Access denied"}, 403
     instance.delete()
-    return
+    return {}, 200
 
 
 def del_hydro(uid, hydro_id):
     instance = HydroBillField.objects.get(id=hydro_id, user_id=uid)
     if instance is None:
         return {"error": "not found"}, 404
+    if instance.user_id != uid:
+        return {"error": "Access denied"}, 403
     instance.delete()
+    return {}, 200
 
 
 def del_commute(uid, commute_id):
@@ -23,3 +28,4 @@ def del_commute(uid, commute_id):
 
     instance = Commute.objects.get(commute_id=commute_id)
     instance.delete()
+    return {}, 200
